requirements.txt
aniso8601==1.2.1
asn1crypto==0.22.0
cffi==1.10.0
click==6.7
cryptography==2.0
enum34==1.1.6
Flask==0.12.2
flask-restplus==0.10.1
functools32==3.2.3.post2
idna==2.5
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.9.6
jsonschema==2.6.0
MarkupSafe==1.0
pycparser==2.18
PyJWT==1.5.2
python-dateutil==2.6.1
pytz==2017.2
six==1.10.0
Werkzeug==0.12.2
